Bullets have a dedicated class now
Player dummy model now rotates properly according to where the mouse is.
Created a rectangle that allows the player to walk around and shoot in.
Bullets now get destroyed when they exit the rectangle and player isn't allowed to leave the rectangle.
Projectiles are now a class and I can change between them and give them unique properties and the like.

Will try to work on creating a hitbox for the player and the bullets, then introduce enemy AI + enemy class to test out collisions and damage. When I do that, I will create a powerup system and see how I can create a "moving stage". Maybe. I don't know maybe I wont have enough time to do all of that and do all the art and animation I want to do